Move events.NewMessage properties to custom.Message
